any other household maitenece yoou can think of
maybe have this device fix plumbing and unclog toilets and recycle human hair and whatever else. 
Please don't make it take over my home. I just want it to be a device. 
It needs to have safety features because i plan on going surfing and letting this thing do it's thing.